Rewrite function that adds book to the library
I made the previous function after adding all of the books to the
libary array and so it was a weird woorkaround to get only the newest book
added to the libary. I had a full array and then was iterating through
it to add the book, but only if it was a new book.

This new function pushes the newBook to the library and then appends
it to the library. Much simpler.